backend/app/services/ai_service.py
# AI Service for Whisper, LLM, and other AI features
import os
from dotenv import load_dotenv

load_dotenv()

class AIService:
    def __init__(self):
        self.openai_api_key = os.getenv("OPENAI_API_KEY")
    
    def transcribe_audio(self, audio_data):
        # Whisper transcription implementation
        pass
    
    def generate_soap_notes(self, consultation_data):
        # LLM SOAP note generation
        pass
    
    def suggest_prescription(self, diagnosis):
        # Prescription auto-completion
        pass



