backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
alembic==1.12.1
pydantic==2.5.0
pydantic-settings==2.1.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
redis==5.0.1
websockets==12.0
openai==1.3.5
whisper==1.1.10
torch==2.1.0
torchaudio==2.1.0
reportlab==4.0.7
python-dotenv==1.0.0
httpx==0.25.2
aiofiles==23.2.1
